requirements.txt
tkinter
google-generativeai
ttkbootstrap